packages/imagen/pyproject.toml
[project]
name = "imagen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "playwright>=1.49.0",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = { find = { include = ["imagen"] } }
